Bug-fixes and 64-bit compatibility

+++ src/BTVBubbleTableView.m
@@ -95,7 +95,7 @@
 	self.bubbleSection = nil;
     
     // Loading new data
-    int count = 0;
+    NSInteger count = 0;
 #if !__has_feature(objc_arc)
     self.bubbleSection = [[[NSMutableArray alloc] init] autorelease];
 #else
@@ -170,7 +170,7 @@
 
 - (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView
 {
-    int result = [self.bubbleSection count];
+    NSInteger result = [self.bubbleSection count];
     if (self.typingBubble != BTVBubbleTypingTypeNobody) result++;
     return result;
 }
@@ -183,7 +183,7 @@
     return [[self.bubbleSection objectAtIndex:section] count] + 1;
 }
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     // Now typing
 	if (indexPath.section >= [self.bubbleSection count])
